Separate  functions and added method to handle larger images
Added an option to insert a prefix in images when downloading them and the possibility to accept or not download from the same target azure environment.
Changed download method for urllib to handle larger images that were previously correpted, but keeping requests as a secondary option to not deal with urllib's encoder.
Added requests timeout to prevent the program from crashing in images. A csv output was also included to store the possible errors captured by the program during execution.